# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Nexus Pathways AI Phase 1 foundation API
 * OpenAPI spec version: 0.1.0
 */
export interface Error {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface ScopeRef {
  id: string;
  name: string;
}

export type ScopeContextLevel = typeof ScopeContextLevel[keyof typeof ScopeContextLevel];


export const ScopeContextLevel = {
  organization: 'organization',
  agency: 'agency',
  region: 'region',
  facility: 'facility',
  program: 'program',
  cohort: 'cohort',
} as const;

export interface ScopeContext {
  organization: ScopeRef;
  agency: ScopeRef | null;
  region: ScopeRef | null;
  facility: ScopeRef | null;
  program: ScopeRef | null;
  cohort: ScopeRef | null;
  level: ScopeContextLevel;
}

export interface CurrentUser {
  id: string;
  displayName: string;
  /** @nullable */
  email: string | null;
  role: string;
  permissions: string[];
  scope: ScopeContext;
}

export type DashboardPortal = typeof DashboardPortal[keyof typeof DashboardPortal];


export const DashboardPortal = {
  learner: 'learner',
  educator: 'educator',
  administrator: 'administrator',
} as const;

export interface Metric {
  label: string;
  value: string;
  detail: string;
}

export interface Dashboard {
  portal: DashboardPortal;
  headline: string;
  scope: ScopeContext;
  metrics: Metric[];
  capabilities: string[];
}

export interface AdminOverview {
  organization: ScopeRef;
  agencies: number;
  regions: number;
  facilities: number;
  programs: number;
  cohorts: number;
  learners: number;
  educators: number;
  syntheticDataNotice: string;
}

export interface AuditEvent {
  id: string;
  action: string;
  category: string;
  actorDisplayName: string;
  /** @nullable */
  resourceType: string | null;
  createdAt: string;
  outcome: string;
}

export interface Facility {
  id: string;
  regionId: string;
  name: string;
}

export interface FacilityInput {
  regionId: string;
  /** @minLength 1 */
  name: string;
}

export interface FacilityUpdate {
  /** @minLength 1 */
  name: string;
}

export interface Program {
  id: string;
  facilityId: string;
  name: string;
}

export interface ProgramInput {
  facilityId: string;
  /** @minLength 1 */
  name: string;
}

export interface ProgramUpdate {
  /** @minLength 1 */
  name: string;
}

export type HierarchyProgram = Program;

export type HierarchyFacility = Facility & {
  programs: HierarchyProgram[];
};

export interface HierarchyRegion {
  id: string;
  agencyId: string;
  name: string;
  facilities: HierarchyFacility[];
}

export interface HierarchyAgency {
  id: string;
  organizationId: string;
  name: string;
  regions: HierarchyRegion[];
}

export interface OrganizationHierarchy {
  organization: ScopeRef;
  agencies: HierarchyAgency[];
}

export type TenantConfigurationModules = {
  abe: boolean;
  hse: boolean;
  specialEducation: boolean;
  accessibility: boolean;
  aiTutor: boolean;
  career: boolean;
  reentry: boolean;
  passport: boolean;
  offlineMode: boolean;
};

export type TenantConfigurationPolicies = {
  /**
     * @minimum 5
     * @maximum 480
     */
  sessionTimeoutMinutes: number;
  /**
     * @minimum 5
     * @maximum 240
     */
  inactivityTimeoutMinutes: number;
};

export interface TenantConfiguration {
  organizationId: string;
  modules: TenantConfigurationModules;
  policies: TenantConfigurationPolicies;
}

export type TenantConfigurationUpdateModules = {
  abe: boolean;
  hse: boolean;
  specialEducation: boolean;
  accessibility: boolean;
  aiTutor: boolean;
  career: boolean;
  reentry: boolean;
  passport: boolean;
  offlineMode: boolean;
};

export type TenantConfigurationUpdatePolicies = {
  /**
     * @minimum 5
     * @maximum 480
     */
  sessionTimeoutMinutes: number;
  /**
     * @minimum 5
     * @maximum 240
     */
  inactivityTimeoutMinutes: number;
};

export interface TenantConfigurationUpdate {
  modules: TenantConfigurationUpdateModules;
  policies: TenantConfigurationUpdatePolicies;
}

export type LearnerCourseworkStatus = typeof LearnerCourseworkStatus[keyof typeof LearnerCourseworkStatus];


export const LearnerCourseworkStatus = {
  assigned: 'assigned',
  completed: 'completed',
} as const;

export interface LearnerCoursework {
  id: string;
  title: string;
  description: string;
  /** @minimum 0 */
  instructionalHours: number;
  status: LearnerCourseworkStatus;
  /** @nullable */
  completedAt: string | null;
}

export interface LearnerResource {
  id: string;
  title: string;
  summary: string;
  content: string;
  /** @pattern ^/learner/resources/[a-z0-9-]+$ */
  route: string;
}

export interface LearnerGoals {
  /**
     * @maxItems 5
     * @items.minLength 1
     * @items.maxLength 280
     */
  goals: string[];
}

export interface LearnerGoalsUpdate {
  /**
     * @maxItems 5
     * @items.minLength 1
     * @items.maxLength 280
     */
  goals: string[];
}

export type PresentationPreferencesTextSize = typeof PresentationPreferencesTextSize[keyof typeof PresentationPreferencesTextSize];


export const PresentationPreferencesTextSize = {
  standard: 'standard',
  large: 'large',
  'extra-large': 'extra-large',
} as const;

export interface PresentationPreferences {
  textSize: PresentationPreferencesTextSize;
  highContrast: boolean;
  reduceMotion: boolean;
}

export type PresentationPreferencesUpdateTextSize = typeof PresentationPreferencesUpdateTextSize[keyof typeof PresentationPreferencesUpdateTextSize];


export const PresentationPreferencesUpdateTextSize = {
  standard: 'standard',
  large: 'large',
  'extra-large': 'extra-large',
} as const;

export interface PresentationPreferencesUpdate {
  textSize: PresentationPreferencesUpdateTextSize;
  highContrast: boolean;
  reduceMotion: boolean;
}

export interface LearnerActivity {
  lastActiveAt: string;
  sessionTimeoutMinutes: number;
  inactivityTimeoutMinutes: number;
}

export interface LearnerHome {
  coursework: LearnerCoursework[];
  goals: LearnerGoals;
  presentationPreferences: PresentationPreferences;
  resources: LearnerResource[];
  /** @minimum 0 */
  instructionalHoursCompleted: number;
  activity: LearnerActivity;
}

/**
 * Request validation failed
 */
export type BadRequestResponse = Error;

/**
 * Authentication is required
 */
export type UnauthorizedResponse = Error;

/**
 * The authenticated user is not authorized for this scope
 */
export type ForbiddenResponse = Error;

/**
 * Resource not found in the authorized scope
 */
export type NotFoundResponse = Error;

export type ListAuditEventsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};