# lib/api-zod/src/generated/types/learnerActivity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Nexus Pathways AI Phase 1 foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface LearnerActivity {
  lastActiveAt: Date;
  sessionTimeoutMinutes: number;
  inactivityTimeoutMinutes: number;
}