# lib/api-zod/src/generated/types/learnerResource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Nexus Pathways AI Phase 1 foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface LearnerResource {
  id: string;
  title: string;
  summary: string;
  content: string;
  /** @pattern ^/learner/resources/[a-z0-9-]+$ */
  route: string;
}