# lib/api-zod/src/generated/types/presentationPreferences.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Nexus Pathways AI Phase 1 foundation API
 * OpenAPI spec version: 0.1.0
 */
import type { PresentationPreferencesTextSize } from './presentationPreferencesTextSize';

export interface PresentationPreferences {
  textSize: PresentationPreferencesTextSize;
  highContrast: boolean;
  reduceMotion: boolean;
}